requirements.txt
Cython==0.29
Kivy==1.10.1
numpy
requests
opencv-python
websocket-client
